ner/pyproject.toml
[tool.poetry]
name = "ner-solution"
version = "0.1.0"
description = ""
authors = ["Pedro Torruella <5025399+ptorru@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.38.0"
openai = "^1.42.0"
llama-index-core = "^0.11.2"
llama-parse = "^0.5.0"
streamlit-code-editor = "^0.1.21"
firecrawl-py = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
